+++ Sources/Sentry/include/SentryHub+Private.h
@@ -59,6 +59,7 @@ NS_ASSUME_NONNULL_BEGIN
 - (void)captureTransaction:(SentryTransaction *)transaction
                   withScope:(SentryScope *)scope
     additionalEnvelopeItems:(NSArray<SentryEnvelopeItem *> *)additionalEnvelopeItems;
+- (void)saveCrashTransaction:(SentryTransaction *)transaction;
 
 - (void)storeEnvelope:(SentryEnvelope *)envelope;
 - (void)captureEnvelope:(SentryEnvelope *)envelope;